Winch: Remove specific scratch registers from ABI trait
